requirements.dev.txt
pytest==7.1.*
pytest-aiohttp==1.0.*
SQLAlchemy-Utils==0.38.*
faker==14.2.*